Updated activity 101
