--- Chapter01/Activities/Activity103/products.csv
@@ -1,1 +1,0 @@
-        {
